Move feature related function to feature utils
