+++ app/assets/scripts/utils/features.js
@@ -1,3 +1,5 @@
+import GJV from 'geojson-validation';
+
 export function geometryToFeature (results) {
   if (results.length > 1) {
     if (results[0].geometry) {
@@ -53,5 +55,41 @@ export function combineFeatureResults (results) {
           'properties': { }
         };
       })
+  });
+}
+
+export function validateGeoJSONPolygon (geojson, cb) {
+  let isFc = true;
+  let isF = true;
+
+  GJV.isFeatureCollection(geojson, (valid, errs) => {
+    if (!valid) isFc = false;
+
+    GJV.isFeature(geojson, (valid, errs) => {
+      if (!valid) isF = false;
+
+      if (!isFc && !isF) {
+        console.log('errors', errs);
+        return cb('Invalid format, provide a Feature or Feature Collection');
+      }
+
+      if (isFc && geojson.features.length > 1) {
+        return cb('Feature Collection should only have 1 feature');
+      }
+
+      let feature = isFc ? geojson.features[0] : geojson;
+
+      if (!feature.geometry) {
+        return cb('Invalid feature provided');
+      }
+
+      GJV.isPolygon(feature.geometry, (valid, errs) => {
+        if (!valid) {
+          return cb('Please provide a feature with a valid polygon');
+        } else {
+          return cb(null, feature.geometry.coordinates[0]);
+        }
+      });
+    });
   });
 }
